requirements.txt
apache-airflow==2.10.2
requests==2.32.3
tenacity==8.5.0
python-dotenv==1.0.1

boto3==1.35.88

snowflake-connector-python==3.15.0
dbt-snowflake==1.8.4

pytest==8.3.4
